pyproject.toml
[tool.poetry]
name = "repl_python3_BreenBot-Discord-Bot"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
chatbot-py = "^2.74"
chatterbot-voice = "^0.1.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
